\name{BaseballStats-package}
\alias{BaseballStats-package}
\alias{BaseballStats}
\docType{package}
\title{
\packageTitle{BaseballStats}
}
\description{
\packageDescription{BaseballStats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BaseballStats}
\packageIndices{BaseballStats}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BaseballStats}

Maintainer: \packageMaintainer{BaseballStats}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
