% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compute_average}
\alias{compute_average}
\title{A function for computing a batter's average.}
\usage{
compute_average(num_hits, num_at_bats)
}
\arguments{
\item{num_hits}{Number of hits}

\item{num_at_bats}{Number of at bats}
}
\description{
This function takes in the number of at-bats and number of hits and will return an average.
}
\examples{
compute_average(10, 50)
}
